[STEP3] 로또 (2등) (#1084)
* refactor: Exception 사유 메세지 추가

* refactor: static block을 사용하여 LottoNumberAutoGenerator 초기화

* docs: STEP3 - 로또(2등) 기능목록 작성

* feat: WinningNumber에 bonusNumber 추가

* feat: 보너스넘버에 따른 등수 조정

* feat: 보너스볼 입력 및 변경된 순위 출력하도록 추가

* refactor: 당첨결과에 따른 등수 비교 method 내 비교문 method 로 분리 및 테스트 케이스 추가

+++ src/main/java/lotto/domain/LottoRunner.java
@@ -7,7 +7,7 @@ public class LottoRunner {
 
 	private static final int PRICE_PER_LOTTO_NUMBER = 1000;
 
-	public LottoRunner(int payment) {
+	public LottoRunner(long payment) {
 		this.lottoMatchResults = new LottoMatchResults(payment);
 		this.lottoNumbers = LottoNumbersGenerator.generate(lottoTotalAmount(payment));
 	}
@@ -16,7 +16,7 @@ public class LottoRunner {
 		return this.lottoNumbers.match(this.lottoMatchResults, winningNumber);
 	}
 
-	public int lottoTotalAmount(int payment) {
+	public long lottoTotalAmount(long payment) {
 		return payment / PRICE_PER_LOTTO_NUMBER;
 	}
